wip(engine): 清理硬编码凭据与不可复现的构建输入
- voice/assets/html/settings.html: 移除随源码公开的 SiliconFlow API key，改为空值 + 提示自填
- tests: Boost/vcpkg 改为按环境变量解析，删除与真实构建漂移的 config_files 模板
- voice models: 钉死 Hugging Face revision 并按 SHA256 校验，记录 silero_vad.onnx 来源
- dictionary/requirements.txt: 加 hash 并在 CI 启用 --require-hashes
- 修 3 条 CodeQL high: float 平方先提升到 double

+++ voice/platforms/windows/src/main.cpp
@@ -311,7 +311,9 @@ int main()
                 double sum_sq = 0.0;
                 for (size_t i = 0; i < count; ++i)
                 {
-                    sum_sq += data[i] * data[i];
+                    // Widen before squaring. float * float is evaluated at float precision and only then
+                    // converted, which throws away the accuracy the double accumulator exists to keep.
+                    sum_sq += static_cast<double>(data[i]) * static_cast<double>(data[i]);
                 }
                 const float rms = count > 0 ? static_cast<float>(std::sqrt(sum_sq / static_cast<double>(count))) : 0.0f;
                 wave_overlay.set_input_level(std::min(1.0f, rms * 8.0f));
@@ -345,7 +347,9 @@ int main()
                 double sum_sq = 0.0;
                 for (size_t i = 0; i < count; ++i)
                 {
-                    sum_sq += data[i] * data[i];
+                    // Widen before squaring. float * float is evaluated at float precision and only then
+                    // converted, which throws away the accuracy the double accumulator exists to keep.
+                    sum_sq += static_cast<double>(data[i]) * static_cast<double>(data[i]);
                 }
                 const float rms = count > 0 ? static_cast<float>(std::sqrt(sum_sq / static_cast<double>(count))) : 0.0f;
                 wave_overlay.set_input_level(std::min(1.0f, rms * 8.0f));
